add more code order

+++ domain/src/main/scala/order/OrderRepository.scala
@@ -13,7 +13,7 @@ class OrderRepository @Inject() (orderDAO: OrderDAO) {
     dao.store(entity2Record(newItem)).map(record2Entity)
   }
 
-  def findOrderDrafByUserId(userId: Int) : Try[Seq[Order]] = Try{
+  def findOrderDrafByUserId(userId: Long): Try[Seq[Order]] = Try {
     dao.findOrderDraftByUserId(userId).get.map(record2Entity)
   }
